- #21: loops are now floats

+++ blender_25/io_xplane2blender/xplane_props.py
@@ -45,10 +45,10 @@ class XPlaneDataref(bpy.types.PropertyGroup):
                                     description="Value",
                                     default=0)
 
-    loop = bpy.props.IntProperty(attr="loop",
+    loop = bpy.props.FloatProperty(attr="loop",
                                 name="Loop Amount",
                                 description="Loop amount of animation, usefull for ever increasing Datarefs. A value of 0 will ignore this setting.",
-                                min=0)
+                                min=0.0)
 
     anim_type = bpy.props.EnumProperty(attr="anim_type",
                                         name="Animation Type",
